Remove datetime bicycle, change service to resource
